requirements.txt
# Core ML/DL frameworks
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
numpy==2.2.6

# Diffusion models and training
diffusers==0.36.0
transformers==4.57.3
accelerate==1.12.0
peft==0.18.0
safetensors==0.7.0
bitsandbytes==0.49.0

# Image processing
Pillow==12.0.0
opencv-python-headless==4.12.0.88
imageio==2.37.2
scikit-image==0.25.2

# Data handling
datasets==4.4.2
pandas==2.3.3
huggingface-hub==0.36.0

# Utilities
tqdm==4.67.1
matplotlib==3.10.8
requests==2.32.5
PyYAML==6.0.3

# API
fastapi==0.115.5
uvicorn==0.32.1
pymongo==4.10.1
